# openapi.yaml
openapi: 3.0.3
info:
  title: RAG Service
  version: 2.1.0
paths:
...
  /private_dbs:
    post:
      tags:
      - private-db
      summary: Create Private Db
      operationId: create_private_db_private_dbs_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivateDBCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDBInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - private-db
      summary: List Private Dbs
      operationId: list_private_dbs_private_dbs_get
      parameters:
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: owner_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Wallet Id
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDBList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /private_dbs/{private_db_id}:
    get:
      tags:
      - private-db
      summary: Get Private Db
      operationId: get_private_db_private_dbs__private_db_id__get
      parameters:
      - name: private_db_id
        in: path
        required: true
        schema:
          type: string
          title: Private Db Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDBInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /private_dbs/{private_db_id}/bind:
    post:
      tags:
      - private-db
      summary: Bind Private Db
      operationId: bind_private_db_private_dbs__private_db_id__bind_post
      parameters:
      - name: private_db_id
        in: path
        required: true
        schema:
          type: string
          title: Private Db Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivateDBBindRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDBBindResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /private_dbs/{private_db_id}/sessions:
    get:
      tags:
      - private-db
      summary: List Private Db Sessions
      operationId: list_private_db_sessions_private_dbs__private_db_id__sessions_get
      parameters:
      - name: private_db_id
        in: path
        required: true
        schema:
          type: string
          title: Private Db Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDBSessionList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /private_dbs/{private_db_id}/sessions/{session_id}:
    delete:
      tags:
      - private-db
      summary: Unbind Private Db Session
      operationId: unbind_private_db_session_private_dbs__private_db_id__sessions__session_id__delete
      parameters:
      - name: private_db_id
        in: path
        required: true
        schema:
          type: string
          title: Private Db Id
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDBUnbindResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PrivateDBBindRequest:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
          description: 操作钱包 ID（兼容字段；推荐使用 Authorization）
        app_id:
          type: string
          title: App Id
          description: 业务插件 ID
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
          description: 可选：数据归属钱包 ID（业务用户）。默认等于 wallet_id。
        session_ids:
          items:
            type: string
          type: array
          title: Session Ids
          description: 绑定的 session_id 列表
      type: object
      required:
      - app_id
      - session_ids
      title: PrivateDBBindRequest
    PrivateDBBindResponse:
      properties:
        private_db_id:
          type: string
          title: Private Db Id
        session_ids:
          items:
            type: string
          type: array
          title: Session Ids
        bound_count:
          type: integer
          title: Bound Count
      type: object
      required:
      - private_db_id
      - session_ids
      - bound_count
      title: PrivateDBBindResponse
    PrivateDBCreateRequest:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
          description: 操作钱包 ID（兼容字段；推荐使用 Authorization）
        app_id:
          type: string
          title: App Id
          description: 业务插件 ID
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
          description: 可选：数据归属钱包 ID（业务用户）。默认等于 wallet_id。
        private_db_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Private Db Id
          description: 可选：自定义私有库 ID
      type: object
      required:
      - app_id
      title: PrivateDBCreateRequest
    PrivateDBInfo:
      properties:
        private_db_id:
          type: string
          title: Private Db Id
        app_id:
          type: string
          title: App Id
        owner_wallet_id:
          type: string
          title: Owner Wallet Id
        status:
          type: string
          title: Status
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
      type: object
      required:
      - private_db_id
      - app_id
      - owner_wallet_id
      - status
      title: PrivateDBInfo
    PrivateDBList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PrivateDBInfo'
          type: array
          title: Items
      type: object
      title: PrivateDBList
    PrivateDBSessionInfo:
      properties:
        session_id:
          type: string
          title: Session Id
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
      type: object
      required:
      - session_id
      title: PrivateDBSessionInfo
    PrivateDBSessionList:
      properties:
        private_db_id:
          type: string
          title: Private Db Id
        app_id:
          type: string
          title: App Id
        owner_wallet_id:
          type: string
          title: Owner Wallet Id
        sessions:
          items:
            $ref: '#/components/schemas/PrivateDBSessionInfo'
          type: array
          title: Sessions
      type: object
      required:
      - private_db_id
      - app_id
      - owner_wallet_id
      title: PrivateDBSessionList
    PrivateDBUnbindResponse:
      properties:
        private_db_id:
          type: string
          title: Private Db Id
        session_id:
          type: string
          title: Session Id
        removed_count:
          type: integer
          title: Removed Count
      type: object
      required:
      - private_db_id
      - session_id
      - removed_count
      title: PrivateDBUnbindResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...